+++ src/main/java/gov/usgs/volcanoes/swarm/OptionsDialog.java
@@ -42,6 +42,7 @@ public class OptionsDialog extends SwarmModalDialog {
   private JTextField pVelocity;
   private JTextField velocityRatio;
   private JCheckBox useLargeCursor;
+  private JCheckBox hideStaleChannel;
 
   private JCheckBox tzInstrument;
   private JRadioButton tzLocal;
@@ -74,6 +75,7 @@ public class OptionsDialog extends SwarmModalDialog {
     pVelocity = new JTextField();
     velocityRatio = new JTextField();
     useLargeCursor = new JCheckBox("Large Helicorder Cursor");
+    hideStaleChannel = new JCheckBox("Hide stale channels");
     tzInstrument = new JCheckBox("Use instrument time zone if available");
     tzLocal = new JRadioButton("Use local machine time zone:");
     tzSpecific = new JRadioButton("Use specific time zone:");
@@ -176,6 +178,7 @@ public class OptionsDialog extends SwarmModalDialog {
     });
     builder.appendSeparator("Other");
     builder.append(useLargeCursor, 7);
+    builder.append(hideStaleChannel, 7);
     builder.nextLine();
 
     dialogPanel = builder.getPanel();
@@ -201,6 +204,7 @@ public class OptionsDialog extends SwarmModalDialog {
    */
   public void setCurrentValues() {
     useLargeCursor.setSelected(swarmConfig.useLargeCursor);
+    hideStaleChannel.setSelected(swarmConfig.hideStaleChannel);
     durationA.setText(Double.toString(swarmConfig.durationA));
     durationB.setText(Double.toString(swarmConfig.durationB));
     durationEnabled.setSelected(swarmConfig.durationEnabled);
@@ -263,6 +267,7 @@ public class OptionsDialog extends SwarmModalDialog {
    */
   public void wasOk() {
     swarmConfig.useLargeCursor = useLargeCursor.isSelected();
+    swarmConfig.hideStaleChannel = hideStaleChannel.isSelected();
     swarmConfig.durationEnabled = durationEnabled.isSelected();
     swarmConfig.durationA = Double.parseDouble(durationA.getText().trim());
     swarmConfig.durationB = Double.parseDouble(durationB.getText().trim());